Refactor code to modules
