+++ src/create-game.js
@@ -1,0 +1,14 @@
+// CREATE A NEW GAME
+fetch("https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/", {
+  method: "POST",
+  headers: {
+    "Content-Type": "application/json",
+  },
+  body: JSON.stringify({
+    name: "Squid Game",
+  }),
+})
+  .then((res) => {
+    return res.json();
+  })
+  .then((data) => console.log(data.result));
